% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_param.R
\name{add_param}
\alias{add_param}
\title{add_param}
\usage{
add_param(ap.newparam, ap.metadata, ap.mode = F, ap.bk = normalizePath("~"))
}
\arguments{
\item{ap.newparam}{Character. Indicates the filename (csv) of the new parameter data.
It has to contain at least three fields: accession_id, accession_name, phenotype_value}

\item{ap.metadata}{Character. Indicates the metadata filename (csv).}

\item{ap.mode}{Flag. F (default) for phenotypical parameter, T for environmental}

\item{ap.bk}{Character. Path to the backup folder. (default home folder).}
}
\value{
NULL. It automatically writes a backup.
}
\description{
The function adds a new parameter to the dataset and performs a backup. It requires
two files: the first containing the actual new parameter data. Inside the file three
fields are mandatory: "accession_id", "accession_name", "phenotype_value". The second
file reports all the metadata information. (Note that environmental and phenotypical
parameters have different sets of metadata. They need to be stored with the correct
number of fields and in the correct order). Both the files are in .csv format.
}
\details{
After the data file has been read it's fileterd so that only the ecotypes among our
selection of 1131 are retained; then the new data are saved in the exaustive
corresponding table (phenotypical or environmental).

Finally the complete dataset is backed up and two sets of tables are printed: one tagged
by the current date, to keep track of history of modifications and a second set tagged as
"current" for the actual use and reference.
}
\examples{

# Example with environmental data
fl = system.file("extdata", "example_add_param.csv", package = "daphneg2")
mt = system.file("extdata", "example_add_param_meta.csv", package = "daphneg2")
add_param(fl, mt, TRUE)

}
