% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arapheno_example}
\alias{arapheno_example}
\title{arapheno_example}
\format{
An object of type \code{data.frame}
}
\usage{
arapheno_example
}
\description{
arapheno_example detailed description2
}
\keyword{datasets}
