% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{environmental}
\alias{environmental}
\title{environmental}
\format{
An object of type \code{data.frame}
}
\usage{
environmental
}
\description{
environmental detailed description2
}
\keyword{datasets}
