% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gemma_annot_example}
\alias{gemma_annot_example}
\title{gemma_annot_example}
\format{
An object of type \code{data.frame}
}
\usage{
gemma_annot_example
}
\description{
gemma_annot_example detailed description2
}
\keyword{datasets}
