% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geno_example}
\alias{geno_example}
\title{geno_example}
\format{
An object of type \code{data.frame}
}
\usage{
geno_example
}
\description{
geno_example detailed description2
}
\keyword{datasets}
