% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_distribution.R
\name{hist_distr}
\alias{hist_distr}
\title{hist_distr}
\usage{
hist_distr(h.par, h.pw = normalizePath("~"))
}
\arguments{
\item{h.par}{character{1}. Parameter name, used as a prefix in the files of interest.}

\item{h.pw}{character{1}. Output path. (default home folder).}
}
\value{
NULL. It automatically produces a plot.
}
\description{
it plots the histogram distribution corresponding to the requested parameter.
}
