% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infos.R
\name{infos}
\alias{infos}
\title{infos}
\usage{
infos(i.input)
}
\arguments{
\item{i.input}{Character. Parameter to print.}
}
\value{
Detailed information for the selected environmental/phenotypical parameter
}
\description{
The function belongs to the dataset management part. It search a parameter of interest
(either phenotypical or environmental) against all the parameters already present in the
collection. If the parameter is there the function returns all the information and metadata
about this parameter. Otherwise it prints an error message and the user should first
use the add_param function and then try again.
}
\examples{
# Checking info for an environmental parameter.
infos("CO_Spring")

# Checking info for a phenotypical parameter.
infos("87_LN16")


}
