% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattan.R
\name{manhattan}
\alias{manhattan}
\title{manhattan}
\usage{
manhattan(mh.par, mh.pw = normalizePath("~"))
}
\arguments{
\item{mh.par}{character{1}. Parameter name, used as a prefix in the files of interest.}

\item{mh.pw}{character{1}. Output path. (default home folder).}
}
\value{
NULL. It automatically produces a plot.
}
\description{
produces mnhattan plots
}
