% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phenotypical}
\alias{phenotypical}
\title{phenotypical}
\format{
An object of type \code{data.frame}
}
\usage{
phenotypical
}
\description{
phenotypical detailed description2
}
\keyword{datasets}
