% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{preproc}
\alias{preproc}
\title{preproc}
\usage{
preproc(
  pp.par,
  pp.option,
  pp.geno,
  pp.cvt = NULL,
  pp.loco = F,
  pp.pw = normalizePath("~")
)
}
\arguments{
\item{pp.par}{character. Name of the parameter of interest.}

\item{pp.option}{character. Label for which kind of analysis is needed (P, E, PxP, ExE, PxE)}

\item{pp.geno}{character. Rdata file with genotype information.}

\item{pp.cvt}{character. Name of the covariate. (default NULL)}

\item{pp.loco}{boolean. Leave one chromosome out approach. (default FALSE).}

\item{pp.pw}{character. Output path. (default home folder).}
}
\value{
NULL. It strarts the preprocessing prodedure and it calls either "univariate" or
"multivariate" in order to produce GWAS files.
}
\description{
The function proprocesses data for GWAS analysis. It distinguishes among five different
modes of preprocessing. Two of them target the univariate analysis: "E" for environmental
target parameter and "P" if the target is phenotypical.
THe latter three start a multivariate analysis: "PxP" phenotype target and phenotype
covariate, ExE for environmental target and environmental covariate and "PxE" for
phenotypical target and environmental covariate respectively.
}
\examples{
gen = system.file("extdata", "example_geno.RData", package = "daphneg2")
preproc("CO_Spring", "E", gen)
}
