% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq_plot.R
\name{QQplot}
\alias{QQplot}
\title{QQplot}
\usage{
QQplot(qq.par, qq.pw = normalizePath("~"))
}
\arguments{
\item{qq.par}{character{1}. Parameter name, used as a prefix in the files of interest.}

\item{qq.pw}{character{1}. Output path. (default home folder).}
}
\value{
NULL. It automatically produces a plot.
}
\description{
it plots the qq plot corresponding to the requested parameter.
}
