% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synchro.R
\name{synchro}
\alias{synchro}
\title{synchro}
\usage{
synchro(s.mode = T, s.bk = normalizePath("~"), s.version = "current")
}
\arguments{
\item{s.mode}{Flag. F (default) for phenotypical parameter, T for environmental.}

\item{s.bk}{Character. Path to the backup folder. (default home folder).}

\item{s.version}{character. Tells which version should be saved in the .RData file ("current" default)}
}
\value{
NULL. It automatically updates RData files.
}
\description{
It takes care that the system it's synchronized - all the parameters added to csv
files in the backup folder should be present also in the RData file.
Updates the data Rd files if a more recent backup is found.
}
\details{
It's hypothically possible to syncrhonize the system with a version that is previous
to the current one. In this case the parameter s.version has to be specified with
the correct version the user wants, tagged with date and time in the format
yyyy-mm-dd_hh-mm-ss.
}
\examples{
# Example with phenotypical data
synchro()

# Example with environmental data
synchro(s.mode = TRUE)


}
