% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{univariate}
\alias{univariate}
\title{univariate}
\usage{
univariate(uv.par, uv.pheno, genotype, uv.pw = normalizePath("~"))
}
\arguments{
\item{uv.par}{character. Name of the parameter of interest.}

\item{uv.pheno}{data.frame. Either phenotypical or environmental.}

\item{genotype}{data.frame. Table with SNP occurrences among ecotypes.}

\item{uv.pw}{character. Output path. (default home folder).}
}
\value{
NULL. It's called by the preprocessing prodedure and it automatically writes
phenotype and genotype files for GWAS in case of univariate analysis.
}
\description{
Internal intermediate function. Starts the preprocessing procedure for
univariate analysis. It matches available phenotype values for the selected parameter
with the available ecotypes in the dataset, it retrieves the corresponding genotipic
information and writes the files that will be fed to the GWAS analysis software.
(In this case only genotype and phenotype).
}
