% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas.R
\name{gwas}
\alias{gwas}
\title{gwas}
\usage{
gwas(
  gemma.name,
  gw.input,
  gw.cv,
  gw.annot,
  gw.cov = NULL,
  gw.miss = 0.05,
  gw.maf = 0.05,
  gw.pw = normalizePath("~")
)
}
\arguments{
\item{gemma.name}{character{1}. Gemma executable name. Version or system may vary.}

\item{gw.input}{character{1}. Parameter name, used as a prefix in the files of interest.}

\item{gw.cv}{integer{1}. Number of covariates. Can be 0, 1 or 2 (in this case is meant >1).}

\item{gw.annot}{annotation}

\item{gw.cov}{character{1}. Covariate name, can be one specific covariate or "all". (defalut NULL)}

\item{gw.miss}{missingness}

\item{gw.maf}{minor allele frequency}

\item{gw.pw}{character{1}. Output path. (default home folder).}
}
\value{
NULL. It automatically performs the whole GWAS analysis procedure.
}
\description{
it performs a gwas analysis. It collects all the input files: phenotype, genotype and,
in case, covariate. Then through the software GEMMA it calculates the kinship matrix,
and it performs the assosiaction analysis.
Subsequently it adujsts the P-values of the associations, calculating FDRs and since the
genotype table is based on a pruned SNP list (meaning that the SNPs that are inherited
together with a target SNPs are filtered out) it eventually retrieves the lists of SNPs
in the same LD units with the significant SNPs in the result table.
}
